Edited activeSection()
The first section that is at least 15% visible is the active section. The loop ends after the active section is found.
Only the amount a section is visible if the section is active. Because, unless it's flat - the top of a section will always be higher than the bottom of the window.