Animation changes
Changed splash image on home page. Now uses a grid to center and stack images.
Also added height to UFO. The height attribute prevents the menu from glitching. The nav was bouncing when the scroll is positioned close to the top of the UFO image.

Changed navigation animation from a square sliding in to expanding square.

Made the scroll function more generic